+++ mchf-eclipse/f7-files.mak
@@ -122,6 +122,7 @@ drivers/audio/audio_filter.c \
 drivers/audio/audio_management.c \
 drivers/audio/freedv_uhsdr.c \
 drivers/audio/freedv_test_data.c \
+drivers/audio/rtty.c \
 basesw/ovi40/Src/adc.c \
 basesw/ovi40/Src/dac.c \
 basesw/ovi40/Src/dma.c \
