+++ mchf-eclipse/files.mak
@@ -122,6 +122,7 @@ drivers/audio/audio_filter.c \
 drivers/audio/audio_management.c \
 drivers/audio/freedv_uhsdr.c \
 drivers/audio/freedv_test_data.c \
+drivers/audio/rtty.c \
 basesw/mcHF/Src/adc.c \
 basesw/mcHF/Src/dac.c \
 basesw/mcHF/Src/dma.c \
