+++ mchf-eclipse/hardware/uhsdr_board.h
@@ -291,13 +291,16 @@ struct mchf_waterfall
 // encoder one
 typedef enum {
     ENC_ONE_MODE_AUDIO_GAIN	 = 0,
+    ENC_ONE_MODE_RTTY_SPEED,
     ENC_ONE_MODE_ST_GAIN,
+    ENC_ONE_MODE_CMP_LEVEL,
     ENC_ONE_NUM_MODES
 } EncoderOneModes;
 //
 // encoder two
 typedef enum {
     ENC_TWO_MODE_RF_GAIN =		0,
+    ENC_TWO_MODE_RTTY_SHIFT,
     ENC_TWO_MODE_SIG_PROC,
     ENC_TWO_MODE_NR,
     ENC_TWO_MODE_NOTCH_F,
@@ -311,7 +314,8 @@ typedef enum {
 typedef enum {
     ENC_THREE_MODE_RIT =			0,
     ENC_THREE_MODE_CW_SPEED,
-    ENC_THREE_MAX_MODE
+    ENC_THREE_MODE_INPUT_CTRL,
+    ENC_THREE_NUM_MODES
 } EncoderThreeModes;
 
 //
